LayoutTests/platform/glib/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/onpaint-iframe-expected.txt



FAIL onpaint should fire after iframe changes. assert_equals: Without changes, onpaint should not fire. expected 0 but got 1
